Resize_IPP_Extraction
